pyproject.toml
[project]
name = "jupyter-base"
version = "0.1.0"
description = "Jupyter notebooks with a reusable backend Python library"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Jupyter-Base contributors" }]
dependencies = [
    "python-dotenv>=1.0.1",
]

[project.urls]
Repository = "https://github.com/your-org/Jupyter-Base"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jupyter_base"]

[dependency-groups]
dev = [
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "ruff>=0.8.0",
]
test = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
]
notebook = [
    "ipykernel>=6.29.0",
    "jupyterlab>=4.2.0",
]
aws = [
    "boto3>=1.35.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"

[tool.mypy]
python_version = "3.11"
packages = ["jupyter_base"]
mypy_path = "src"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.coverage.run]
source = ["jupyter_base"]
branch = true
relative_files = true
